recipe caching, restructuring
